refactor(proxy): move the profile-routing test harness into tests/fixtures
`integration_inference_profiles.rs` was 593 LOC, well over the ratchet, and
about half of it was harness rather than test: three mock ports, a spawned
proxy with a recording upstream, and four assertion helpers. A second suite
covering the pinned endpoint's default profile is coming, and it needs all
of that — duplicating ~290 lines to get it would be the worse trade.

Split in two along the file's own section markers rather than into one
fixture module, because a single file would have landed at 317 LOC — over
budget on the day it was created. `profile_mocks.rs` holds the ports;
`profile_harness.rs` holds the spawned proxy and the assertions that read
the forwarded body.

The tests themselves are untouched: same names, same assertions, 13 passed
before and 13 after. Only visibility changed, widening the harness items to
`pub(crate)` so the suite can reach them across the module boundary.

Full pre-commit green.
