+++ crates/gglib-proxy/tests/fixtures/mod.rs
@@ -1,4 +1,6 @@
 //! Test fixtures shared by proxy integration tests.
 #![allow(dead_code)]
 pub(crate) mod common;
+pub(crate) mod profile_harness;
+pub(crate) mod profile_mocks;
 pub(crate) mod sse;
