+++ crates/gglib-proxy/tests/integration_inference_profiles.rs
@@ -7,314 +7,20 @@
 //!
 //! The proxy is the real `gglib_proxy::serve` with mock ports, as in
 //! `integration_proxy_pipeline.rs`.
+//!
+//! The mock ports and the spawned-proxy harness live in
+//! `fixtures/profile_mocks.rs` and `fixtures/profile_harness.rs` so that a
+//! second suite can drive the same proxy without duplicating them.
 
-use std::sync::{Arc, Mutex};
-use std::time::Duration;
-
-use async_trait::async_trait;
-use axum::{Json, Router, routing::post};
-use reqwest::Client;
 use serde_json::{Value, json};
-use tokio::net::TcpListener;
-use tokio_util::sync::CancellationToken;
 
-use gglib_core::Settings;
-use gglib_core::domain::{InferenceConfig, InferenceProfile};
-use gglib_core::ports::{
-    CatalogError, ModelCatalogPort, ModelLaunchSpec, ModelRuntimeError, ModelRuntimePort,
-    ModelSummary, RepositoryError, RunningTarget, SettingsRepository,
-};
+use gglib_core::domain::InferenceConfig;
 
 mod fixtures;
-use fixtures::common::make_mcp_service;
-
-const MODEL: &str = "qwen";
-
-// ─── Mock ports ────────────────────────────────────────────────────────────
-
-/// Runtime that always reports the mock upstream as running, and records the
-/// model name it was asked to launch.
-#[derive(Debug)]
-struct RecordingRuntime {
-    port: u16,
-    launched: Arc<Mutex<Vec<String>>>,
-}
-
-#[async_trait]
-impl ModelRuntimePort for RecordingRuntime {
-    async fn admit(
-        &self,
-        model_name: &str,
-        _num_ctx: Option<u64>,
-        _default_ctx: u64,
-        _overrides: gglib_core::ports::LaunchOverrides,
-    ) -> Result<gglib_core::ports::Admission, ModelRuntimeError> {
-        self.launched.lock().unwrap().push(model_name.to_owned());
-        Ok(gglib_core::ports::Admission::detached(
-            RunningTarget::local(self.port, 1, model_name.to_owned(), 4096, false),
-        ))
-    }
-    async fn current_model(&self) -> Option<RunningTarget> {
-        None
-    }
-    async fn stop_current(&self) -> Result<(), ModelRuntimeError> {
-        Ok(())
-    }
-}
-
-/// Catalog resolving an explicit set of names by exact match.
-#[derive(Debug)]
-struct NamedCatalog {
-    names: Vec<String>,
-    /// Per-model stored defaults, returned for every resolved model.
-    inference_defaults: Option<InferenceConfig>,
-}
-
-impl NamedCatalog {
-    fn summary(&self, name: &str) -> ModelSummary {
-        ModelSummary {
-            dialect: None,
-            template_caps: None,
-            id: 1,
-            name: name.to_owned(),
-            tags: Vec::new(),
-            capabilities: gglib_core::domain::ModelCapabilities::empty(),
-            param_count: "7B".into(),
-            quantization: None,
-            architecture: None,
-            created_at: 0,
-            file_size: 0,
-            context_length: None,
-            inference_defaults: self.inference_defaults.clone(),
-            defaults_origin: None,
-            server_defaults: None,
-        }
-    }
-}
-
-#[async_trait]
-impl ModelCatalogPort for NamedCatalog {
-    async fn list_models(&self) -> Result<Vec<ModelSummary>, CatalogError> {
-        Ok(self.names.iter().map(|n| self.summary(n)).collect())
-    }
-    async fn resolve_model(&self, name: &str) -> Result<Option<ModelSummary>, CatalogError> {
-        Ok(self
-            .names
-            .iter()
-            .any(|n| n == name)
-            .then(|| self.summary(name)))
-    }
-    async fn resolve_for_launch(
-        &self,
-        _name: &str,
-    ) -> Result<Option<ModelLaunchSpec>, CatalogError> {
-        Ok(None)
-    }
-}
-
-/// Settings repository serving a fixed profile list.
-struct ProfileSettings {
-    profiles: Vec<InferenceProfile>,
-    trust_client_sampling: bool,
-}
-
-#[async_trait]
-impl SettingsRepository for ProfileSettings {
-    async fn load(&self) -> Result<Settings, RepositoryError> {
-        Ok(Settings {
-            inference_profiles: Some(self.profiles.clone()),
-            trust_client_sampling: Some(self.trust_client_sampling),
-            ..Settings::with_defaults()
-        })
-    }
-    async fn save(&self, _: &Settings) -> Result<(), RepositoryError> {
-        Ok(())
-    }
-}
-
-// ─── Harness ───────────────────────────────────────────────────────────────
-
-/// Everything a test needs to drive one request and inspect what the upstream
-/// saw.
-struct Harness {
-    proxy_url: String,
-    /// Bodies the mock upstream received, in order.
-    forwarded: Arc<Mutex<Vec<Value>>>,
-    /// Model names the runtime was asked to launch.
-    launched: Arc<Mutex<Vec<String>>>,
-    _cancel: CancellationToken,
-}
-
-impl Harness {
-    /// POST a non-streaming chat completion and return the HTTP status.
-    ///
-    /// Non-streaming keeps the assertions about the *request* body free of any
-    /// SSE machinery.
-    async fn post(&self, body: Value) -> reqwest::Response {
-        Client::new()
-            .post(format!("{}/v1/chat/completions", self.proxy_url))
-            .json(&body)
-            .send()
-            .await
-            .expect("request reaches the proxy")
-    }
-
-    /// GET the model list as JSON.
-    async fn models(&self) -> Value {
-        Client::new()
-            .get(format!("{}/v1/models", self.proxy_url))
-            .send()
-            .await
-            .expect("request reaches the proxy")
-            .json()
-            .await
-            .expect("model list is JSON")
-    }
-
-    /// The single body the upstream received. Panics if it saw none.
-    fn only_forwarded(&self) -> Value {
-        let bodies = self.forwarded.lock().unwrap();
-        assert_eq!(bodies.len(), 1, "expected exactly one upstream call");
-        bodies[0].clone()
-    }
-}
-
-async fn spawn(
-    profiles: Vec<InferenceProfile>,
-    catalog_names: &[&str],
-    model_defaults: Option<InferenceConfig>,
-    trust_client_sampling: bool,
-) -> Harness {
-    let forwarded: Arc<Mutex<Vec<Value>>> = Arc::new(Mutex::new(Vec::new()));
-    let cancel = CancellationToken::new();
-
-    // Mock upstream: record the body, return a minimal non-streaming reply.
-    let upstream_listener = TcpListener::bind("127.0.0.1:0").await.expect("bind");
-    let upstream_port = upstream_listener.local_addr().unwrap().port();
-    let recorder = Arc::clone(&forwarded);
-    let app = Router::new().route(
-        "/v1/chat/completions",
-        post(move |Json(body): Json<Value>| {
-            let recorder = Arc::clone(&recorder);
-            async move {
-                recorder.lock().unwrap().push(body);
-                Json(json!({
-                    "id": "chatcmpl-test",
-                    "object": "chat.completion",
-                    "created": 0,
-                    "model": MODEL,
-                    "choices": [{
-                        "index": 0,
-                        "message": {"role": "assistant", "content": "ok"},
-                        "finish_reason": "stop"
-                    }],
-                    "usage": {"prompt_tokens": 1, "completion_tokens": 1, "total_tokens": 2}
-                }))
-            }
-        }),
-    );
-    let upstream_cancel = cancel.clone();
-    tokio::spawn(async move {
-        axum::serve(upstream_listener, app)
-            .with_graceful_shutdown(upstream_cancel.cancelled_owned())
-            .await
-            .ok();
-    });
-
-    // Proxy.
-    let launched: Arc<Mutex<Vec<String>>> = Arc::new(Mutex::new(Vec::new()));
-    let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-    let addr = listener.local_addr().unwrap();
-    let runtime: Arc<dyn ModelRuntimePort> = Arc::new(RecordingRuntime {
-        port: upstream_port,
-        launched: Arc::clone(&launched),
-    });
-    let catalog: Arc<dyn ModelCatalogPort> = Arc::new(NamedCatalog {
-        names: catalog_names.iter().map(|n| (*n).to_owned()).collect(),
-        inference_defaults: model_defaults,
-    });
-    let mcp = make_mcp_service();
-    let proxy_cancel = cancel.clone();
-    tokio::spawn(async move {
-        gglib_proxy::serve(
-            listener,
-            4096,
-            runtime,
-            catalog,
-            mcp,
-            proxy_cancel,
-            Arc::new(ProfileSettings {
-                profiles,
-                trust_client_sampling,
-            }),
-            None, // inference_override
-            false,
-            None,
-            gglib_proxy::slot_eviction::DiskBudget::Auto,
-            std::sync::Arc::new(gglib_core::cache_metrics::CacheMetricsStore::new()),
-            std::sync::Arc::new(gglib_core::domain::defects::ModelDefectLedger::new()),
-            &gglib_core::ProxyAccessConfig::default(),
-        )
-        .await
-        .ok();
-    });
-
-    tokio::time::sleep(Duration::from_millis(50)).await;
-    Harness {
-        proxy_url: format!("http://{addr}"),
-        forwarded,
-        launched,
-        _cancel: cancel,
-    }
-}
-
-/// A sparse `coding` profile: it has exactly one opinion.
-fn coding_profile() -> InferenceProfile {
-    InferenceProfile {
-        name: "coding".to_owned(),
-        description: None,
-        config: InferenceConfig {
-            temperature: Some(0.2),
-            ..Default::default()
-        },
-        list_in_models: false,
-    }
-}
-
-fn chat_request(model: &str) -> Value {
-    json!({
-        "model": model,
-        "stream": false,
-        "messages": [{"role": "user", "content": "hi"}],
-    })
-}
-
-/// Assert a sampling parameter's value.
-///
-/// Compared with a tolerance because these travel as `f32` through
-/// `InferenceConfig` and widen to `f64` in JSON, so an exact match on a literal
-/// like `0.2` fails on the widening artifact rather than on behaviour.
-#[track_caller]
-fn assert_param(body: &Value, key: &str, expected: f64) {
-    let actual = body
-        .get(key)
-        .and_then(Value::as_f64)
-        .unwrap_or_else(|| panic!("{key} missing from forwarded body: {body}"));
-    assert!(
-        (actual - expected).abs() < 1e-6,
-        "{key}: expected {expected}, got {actual}"
-    );
-}
-
-/// Assert a sampling parameter never reached the wire, so llama.cpp's own
-/// default applies. The normal outcome for six of the seven since ADR 0003.
-#[track_caller]
-fn assert_deferred(body: &Value, key: &str) {
-    assert!(
-        body.get(key).is_none(),
-        "{key} must be deferred to llama.cpp, but the forwarded body carries {body}"
-    );
-}
+use fixtures::profile_harness::{
+    assert_deferred, assert_param, chat_request, coding_profile, spawn,
+};
+use fixtures::profile_mocks::MODEL;
 
 // ─── Tests ─────────────────────────────────────────────────────────────────
